backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib[bcrypt]
requests
email-validator
pymysql
python-dotenv
python-multipart
httpx